Silence some warnings.

+++ src/core/files.cpp
@@ -9,7 +9,7 @@ namespace MgCore
         std::ifstream in(filename, std::ios::in | std::ios_base::ate);
         if (in) {
             std::string contents;
-            contents.resize(in.tellg());
+            contents.resize(static_cast<unsigned int>(in.tellg()));
             in.seekg(0, std::ios::beg);
             in.read(&contents[0], contents.size());
             in.close();
@@ -25,7 +25,7 @@ namespace MgCore
 
         if ( vf && vf->open("r") ) {
             std::string contents;
-            contents.resize( vf->size() );
+            contents.resize( static_cast<unsigned int>(vf->size()) );
             vf->read(&contents[0], contents.size());
             vf->close();
             return contents;
